Revert "ER-304 geocoding"
