+++ src/Employer/Employer.Web/Orchestrators/Part1/EmployerOrchestrator.cs
@@ -86,9 +86,7 @@ namespace Esfa.Recruit.Employer.Web.Orchestrators.Part1
                 AddressLine2 = m.AddressLine2,
                 AddressLine3 = m.AddressLine3,
                 AddressLine4 = m.AddressLine4,
-                Postcode = m.Postcode.AsPostcode(),
-                Latitude = null,
-                Longitude = null
+                Postcode = m.Postcode.AsPostcode()
             };
 
             return await ValidateAndExecute(
