+++ src/Employer/Esfa.Recruit.Azure.Resources/azuredeploy.parameters.json
@@ -70,13 +70,6 @@
     },
     "SlackWebhook": {
       "value": ""
-    },
-    "Geocode": {
-      "value": {
-        "GeocodePostcodesIoUrl": "",
-        "GeocodePostcodeAnywhereUrl": "",
-        "GeocodePostcodeAnywhereKey": ""
-      }
     }
   }
 }
