+++ src/Jobs/Recruit.Vacancies.Jobs/appSettings.Development.json
@@ -25,10 +25,5 @@
   },
   "Slack": {
     "WebHookUrl": ""
-  },
-  "Geocode": {
-    "PostcodesIoUrl": "http://api.postcodes.io",
-    "PostcodeAnywhereUrl": "http://services.postcodeanywhere.co.uk",
-    "PostcodeAnywhereKey": ""
   }
 }
