+++ src/Jobs/Recruit.Vacancies.Jobs/appSettings.json
@@ -32,10 +32,5 @@
   },
   "Slack": {
     "WebHookUrl": ""
-  },
-  "Geocode": {
-    "PostcodesIoUrl": "",
-    "PostcodeAnywhereUrl": "",
-    "PostcodeAnywhereKey": ""
-  } 
+  }
 }
