--- src/Shared/Recruit.Vacancies.Client/Application/Commands/GeocodeVacancyCommand.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using Esfa.Recruit.Vacancies.Client.Domain.Messaging;
-using MediatR;
-
-namespace Esfa.Recruit.Vacancies.Client.Application.Commands
-{
-    public class GeocodeVacancyCommand : CommandBase, ICommand, IRequest
-    {
-        public Guid VacancyId { get; set; }
-    }
-}
