+++ src/Shared/Recruit.Vacancies.Client/Infrastructure/Client/IJobsVacancyClient.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Esfa.Recruit.Vacancies.Client.Domain.Entities;
 using Esfa.Recruit.Vacancies.Client.Infrastructure.QueryStore.Projections.EditVacancyInfo;
 using Esfa.Recruit.Vacancies.Client.Infrastructure.QueryStore.Projections.LiveVacancy;
 using Esfa.Recruit.Vacancies.Client.Infrastructure.Services.Models;
@@ -17,6 +16,5 @@ namespace Esfa.Recruit.Vacancies.Client.Infrastructure.Client
         Task CreateVacancyReview(long vacancyReference);
         Task<IEnumerable<LiveVacancy>> GetLiveVacancies();
         Task CloseVacancy(Guid vacancyId);
-        Task EnsureVacancyIsGeocodedAsync(Guid vacancyId);
     }
 }
