--- src/Shared/Recruit.Vacancies.Client/Infrastructure/Services/Geocode/GeocodeConfiguration.cs
@@ -1,9 +1,0 @@
-﻿namespace Esfa.Recruit.Vacancies.Client.Infrastructure.Services.Geocode
-{
-    public class GeocodeConfiguration
-    {
-        public string PostcodesIoUrl { get; set; }
-        public string PostcodeAnywhereUrl { get; set; }
-        public string PostcodeAnywhereKey { get; set; }
-    }
-}
