--- src/Shared/Recruit.Vacancies.Client/Infrastructure/Services/Geocode/IGeocodeService.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace Esfa.Recruit.Vacancies.Client.Infrastructure.Services.Geocode
-{
-    public interface IGeocodeService
-    {
-        Task<Geocode> Geocode(string postcode);
-    }
-}
