--- src/Shared/Recruit.Vacancies.Client/Infrastructure/Services/Geocode/IGeocodeServiceFactory.cs
@@ -1,7 +1,0 @@
-﻿namespace Esfa.Recruit.Vacancies.Client.Infrastructure.Services.Geocode
-{
-    public interface IGeocodeServiceFactory
-    {
-        IGeocodeService GetGeocodeService();
-    }
-}
