+++ src/Shared/Recruit.Vacancies.Client/Ioc/ServiceCollectionExtensions.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using Esfa.Recruit.Vacancies.Client.Application.Configuration;
+﻿using Esfa.Recruit.Vacancies.Client.Application.Configuration;
 using Esfa.Recruit.Vacancies.Client.Application.Events;
 using Esfa.Recruit.Vacancies.Client.Application.CommandHandlers;
 using Esfa.Recruit.Vacancies.Client.Application.Services;
@@ -18,7 +17,6 @@ using Esfa.Recruit.Vacancies.Client.Infrastructure.QueryStore;
 using Esfa.Recruit.Vacancies.Client.Infrastructure.Repositories;
 using Esfa.Recruit.Vacancies.Client.Infrastructure.SequenceStore;
 using Esfa.Recruit.Vacancies.Client.Infrastructure.Services;
-using Esfa.Recruit.Vacancies.Client.Infrastructure.Services.Geocode;
 using Esfa.Recruit.Vacancies.Client.Infrastructure.StorageQueue;
 using FluentValidation;
 using MediatR;
@@ -66,14 +64,11 @@ namespace Microsoft.Extensions.DependencyInjection
             services.AddTransient<IEmployerAccountService, EmployerAccountService>();
             services.AddTransient<IGetMinimumWages, StubNationalMinimumWageService>();
             services.AddTransient<IGenerateVacancyNumbers, MongoSequenceStore>();
-            services.AddTransient<IApprenticeshipProgrammeProvider, ApprenticeshipProgrammeProvider>();
+            services.AddTransient<IApprenticeshipProgrammeProvider, ApprenticeshipProgrammeProvider>(); 
             
             services.Configure<SlackConfiguration>(configuration.GetSection("Slack"));
             services.AddTransient<INotifyVacancyReviewUpdates, SlackNotifyVacancyReviewUpdates>();
             services.AddTransient<ISlackClient, SlackClient>();
-
-            services.Configure<GeocodeConfiguration>(configuration.GetSection("Geocode"));
-            services.AddTransient<IGeocodeServiceFactory, GeocodeServiceFactory>();
         }
 
         private static void AddRepositories(this IServiceCollection services, IConfiguration configuration)
