+++ src/Shared/Recruit.Vacancies.Client/Recruit.Vacancies.Client.csproj
@@ -27,7 +27,6 @@
     <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="2.0.0" />
     <PackageReference Include="MongoDB.Driver" Version="2.5.0" />
     <PackageReference Include="Newtonsoft.Json" Version="10.0.3" />
-    <PackageReference Include="RestSharp" Version="106.2.2" />
     <PackageReference Include="SFA.DAS.Account.Api.Client" Version="1.2.0.62882" />
     <PackageReference Include="WindowsAzure.Storage" Version="9.1.0" />
     <PackageReference Include="FluentValidation" Version="7.5.2" />
